Revert "Move KedroStarterSpec to framework __init__"
This reverts commit 611eedbdf1f807d9094cb5ad5ab86b2059cbecd1.
