+++ features/steps/test_plugin/plugin.py
@@ -1,7 +1,7 @@
 """Dummy plugin with simple hook implementations."""
 import logging
 
-from kedro.framework import KedroStarterSpec
+from kedro.framework.cli.starters import KedroStarterSpec
 from kedro.framework.hooks import hook_impl
 
 logger = logging.getLogger(__name__)
